contain io calls within tsnr file.
